% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.R
\name{ages.undated}
\alias{ages.undated}
\title{Model ages between two dated levels}
\usage{
ages.undated(position, set = get("info"), draw = TRUE)
}
\arguments{
\item{position}{Position of the to-be-estimated undated layer. Should be larger than the layer above but smaller than the layer below it.}

\item{set}{This option reads the 'info' variable, which contains the data and the model output.}

\item{draw}{Whether or not to draw the age distribution.}
}
\description{
Model ages of undated levels, by for each MCMC iteration finding the age of the layer above and of the layer below, and sampling a random age from a uniform distribution between the age estimates of the two ages.
}
