% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{Altitude_totalSCore_plot}
\alias{Altitude_totalSCore_plot}
\title{A scatter plot demonstrating the relationship between average altitude of a country and the total quality score}
\usage{
Altitude_totalSCore_plot(data = coffee)
}
\description{
A scatter plot demonstrating the relationship between average altitude of a country and the total quality score
}
