% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee.R
\name{assign_to_global}
\alias{assign_to_global}
\title{Buck CE, Kenworthy JB, Litton CD, Smith AFM, 1991. Combining archaeological and radiocarbon information: a Bayesian approach to calibration. Antiquity 65, 808-821.
Christen JA, 2003. Bwigg: An Internet facility for Bayesian radiocarbon wiggle-matching. Internet Archaeology 13. \doi{10.11141/ia.13.2}
Christen JA, Fox C 2010. A General Purpose Sampling Algorithm for Continuous Distributions (the t-walk). Bayesian Analysis 5, 263-282. \url{http://ba.stat.cmu.edu/journal/2010/vol05/issue02/christen.pdf}}
\usage{
assign_to_global(key, val, pos = 1)
}
\description{
Buck CE, Kenworthy JB, Litton CD, Smith AFM, 1991. Combining archaeological and radiocarbon information: a Bayesian approach to calibration. Antiquity 65, 808-821.
Christen JA, 2003. Bwigg: An Internet facility for Bayesian radiocarbon wiggle-matching. Internet Archaeology 13. \doi{10.11141/ia.13.2}
Christen JA, Fox C 2010. A General Purpose Sampling Algorithm for Continuous Distributions (the t-walk). Bayesian Analysis 5, 263-282. \url{http://ba.stat.cmu.edu/journal/2010/vol05/issue02/christen.pdf}
}
