% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{avg_location_altitude_totalScore}
\alias{avg_location_altitude_totalScore}
\title{Correlation between altitude and the quality of coffee}
\usage{
avg_location_altitude_totalScore(data = coffee)
}
\description{
this function produces a model that investigates linear relationship between average altitude of a location and the total score for the coffee quality
}
\details{
the data for average altitude and total score was not normal. outliers were deleted and boxcox transformation was applied

The fianl model outputs slope and intercept for predicting square of total score from and logarithm of average altitude
}
\examples{
 avg_location_altitude_totalScore()
}
\keyword{boxcox}
\keyword{linear}
\keyword{regression,}
\keyword{simple}
\keyword{transformation}
