% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{Best_quality}
\alias{Best_quality}
\title{country with the highest quality coffee}
\usage{
Best_quality(data = coffee)
}
\value{
country that produces the highest quality coffee
}
\description{
\code{Best_quality} returns the country that produces the coffee that has the highest total score
}
