% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{bw_contrast}
\alias{bw_contrast}
\title{What colour should you use for text}
\usage{
bw_contrast(x)
}
\description{
Should you use black or white text for the given background colour
}
\references{
https://stackoverflow.com/a/3943023
}
\keyword{internal}
