% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_names.R
\name{cc}
\alias{cc}
\title{Easy quote names}
\usage{
cc(..., .clip = TRUE)
}
\arguments{
\item{...}{Unquoted names (separated by commas) that you wish to quote;
empty arguments (e.g. third item in \verb{one,two,,four}) will be returned as
blanks.}

\item{.clip}{Should the code to generate the constructed character vector be
copied to your system clipboard; defaults to TRUE.}
}
\value{
a character vector
}
\description{
Easy quote names
}
\note{
The system clipboard will only be used in interactive sessions.
Outer whitespace is automatically trimmed.
}
\examples{
cc(dale, audrey, laura, hawk)

}
\author{
Tim Taylor
}
