% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_packages}
\alias{check_packages}
\title{Check Packages}
\usage{
check_packages(names)
}
\arguments{
\item{names}{the names of the packages to check}
}
\description{
This function allows you to check if a package is installed and if it is not, install it. 
If it is installed, then the package is loaded.
}
\examples{
check_packages()
}
\keyword{packages}
