% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classmates}
\alias{classmates}
\title{Title}
\format{
character vector
}
\source{
The mid of George Lucas \url{www.starwars.com}
}
\usage{
classmates
}
\description{
A vector of 10 randomly chosen names from the starwars universe.
A short description
}
\keyword{datasets}
