% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tweets.R
\name{clean_tweets}
\alias{clean_tweets}
\title{Cleaning Tweets}
\usage{
clean_tweets(file)
}
\arguments{
\item{file}{A string that is the name of a .csv file of tweets}
}
\description{
This function allows you to clean tweets given a .csv file of tweets.
}
\examples{
clean_tweets()
}
\keyword{tweets}
