% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee-package.R
\docType{package}
\name{coffee-package}
\alias{coffee}
\alias{coffee-package}
\title{coffee: Chronological Ordering for Fossils and Environmental Events}
\description{
While individual calibrated radiocarbon dates can span several centuries, combining multiple dates together with any chronological constraints can make a chronology much more robust and precise. This package uses Bayesian methods to enforce the chronological ordering of radiocarbon and other dates, for example for trees with multiple radiocarbon dates spaced at exactly known intervals (e.g., 10 annual rings). For methods see Christen 2003 \doi{10.11141/ia.13.2}. Another example is sites where the relative chronological position of the dates is taken into account - the ages of dates further down a site must be older than those of dates further up (Buck, Kenworthy, Litton and Smith 1991 \doi{10.1017/S0003598X00080534}; Nicholls and Jones 2001 \doi{10.1111/1467-9876.00250}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Maarten14C/coffee}
}

}
\author{
\strong{Maintainer}: Maarten Blaauw \email{maarten.blaauw@qub.ac.uk} (\href{https://orcid.org/0000-0002-5680-1515}{ORCID})

Authors:
\itemize{
  \item Marco A. Aquino Lopez \email{aquino@cimat.mx} (\href{https://orcid.org/0000-0002-5076-7205}{ORCID})
}

Other contributors:
\itemize{
  \item J. Andres Christen \email{jac@cimat.mx} (\href{https://orcid.org/0000-0002-5795-4345}{ORCID}) [contributor, copyright holder]
}

}
\keyword{internal}
