% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{CoffeeBeanColor_totalQuality}
\alias{CoffeeBeanColor_totalQuality}
\title{Difference between coffee color bean and total quality grade}
\usage{
CoffeeBeanColor_totalQuality(data = coffee)
}
\description{
this function compares the total score for total coffee quality between different coffee bean color based on Tukey's honestly significant difference test
}
\details{
If the coffee bean color is unknown, it was removed from the study.
}
\keyword{HSD}
\keyword{Tukey}
\keyword{linear}
\keyword{model,}
\keyword{regression}
\keyword{simple}
