% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{colorBean_totalScore_plot}
\alias{colorBean_totalScore_plot}
\title{A barplot of coffee total quality score grouped by the coffee bean color}
\usage{
colorBean_totalScore_plot(data = coffee)
}
\description{
A barplot of coffee total quality score grouped by the coffee bean color
}
