% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{Country_highestProduction_plot}
\alias{Country_highestProduction_plot}
\title{A barplot of countries that are the largest coffee producers in descending order}
\usage{
Country_highestProduction_plot(data = coffee)
}
\description{
A barplot of countries that are the largest coffee producers in descending order
}
