% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq.R
\name{cq}
\alias{cq}
\alias{ccq}
\title{Easy quote characters}
\usage{
cq(...)

ccq(x)
}
\arguments{
\item{...}{Unquoted variables (seperated by commas) that you wish to quote}

\item{x}{Character vector which you wish to split on spaces}
}
\description{
Easy quote characters
}
