% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_page.R
\name{download_page}
\alias{download_page}
\alias{get_dl_history}
\title{Download a webpage}
\usage{
download_page(
  url,
  destination = Sys.getenv("WEBPAGE_DIR"),
  history = Sys.getenv("WEBPAGE_DOWNLOAD_HISTORY")
)

get_dl_history(history = Sys.getenv("WEBPAGE_DOWNLOAD_HISTORY"))
}
\arguments{
\item{url}{Webpage to download with.}

\item{destination}{Folder to store webpage.}

\item{history}{Where to save the download history csv.}
}
\value{
\itemize{
\item download_page()  - result (invisibly) of call to system2; 0 = successful.
\item get_dl_history() - data frame of downloaded webpages.
}
}
\description{
Downloads a single webpage using wget and some hardcoded options.
}
\details{
\code{download_page()} calls wget with the following options:\preformatted{* adjust-extension:  save HTML/CSS documents with proper extensions
* span-hosts:        go to foreign hosts when recursive
* convert-links:     make links in downloaded HTML or CSS point to local files
* backup-converted:  before converting file X, back up as X.orig
* page-requisites:   get all images, etc. needed to display HTML page
}
}
\references{
https://www.gnu.org/software/wget/manual/wget.html#Recursive-Retrieval-Options
}
