% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{get_tweets}
\alias{get_tweets}
\title{Getting Tweets}
\usage{
get_tweets(keyword, num_tweets = 500)
}
\arguments{
\item{keyword}{A string that is the keyword included in tweets}

\item{num_tweets}{The number of tweets to get, default is 500}
}
\description{
This function allows you to save tweets that contain a given keyword in a .csv file.
}
\examples{
get_tweets()
}
\keyword{tweets}
