% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk.R
\name{IAT}
\alias{IAT}
\title{calculate the Integrated Autocorrelation Time}
\usage{
IAT(set, par = 0, from = 1, to)
}
\arguments{
\item{set}{This option reads the 'info' variable, which contains the data and the model output.}

\item{par}{The parameter to test. Defaults to 0.}

\item{from}{The first of the iterations of the MCMC run to check. Defaults to the first one.}

\item{to}{The last of the iterations of the MCMC run to check. Defaults to the last one.}
}
\value{
The IAT value
}
\description{
Calculate the Integrated Autocorrelation Time, which gives the proposed value for thinning. E.g., if the IAT is 80, it is good to thin the MCMC run by storing only every 80 iterations. This method is slower than GetAutoCov, but much better.
}
\author{
Andres Christen
}
