% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{libraries}
\alias{libraries}
\title{Coffee Preferences Shiny App}
\description{
This app allows users to explore coffee consumption patterns based on age, gender,
and favorite coffee preferences. Users can filter data by age group, gender,
and favorite coffee type and visualize the results using bar plots.
}
