% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bar_graph.R
\name{make_bar_graph}
\alias{make_bar_graph}
\title{Making a Bar Graph}
\usage{
make_bar_graph(file)
}
\arguments{
\item{file}{A .csv file as a string of clean tweets}
}
\description{
This function allows you to make a bar graph given a .csv file of clean tweets.
}
\examples{
make_bar_graph()
}
\keyword{bargraph}
