% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cat_graph.R
\name{make_cat_graph}
\alias{make_cat_graph}
\title{Making a Category Graph for Emotions}
\usage{
make_cat_graph(file)
}
\arguments{
\item{file}{A .csv file}
}
\description{
This function allows you to make a bar graph given a .csv file.
}
\examples{
make_cat_graph()

}
\keyword{category}
\keyword{graph}
