% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_groups}
\alias{make_groups}
\title{Generate pair combination of names}
\usage{
make_groups(names)
}
\arguments{
\item{names}{name of individuals}
}
\value{
matrix of names pair combinations
}
\description{
Generate pair combination of names
}
