% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_groups_three}
\alias{make_groups_three}
\title{Generate triple combination of names}
\usage{
make_groups_three(names)
}
\arguments{
\item{names}{name of individuals}
}
\value{
matrix of names triple combinations
}
\description{
Generate triple combination of names
}
