% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wordcloud.R
\name{make_wordcloud}
\alias{make_wordcloud}
\title{Making a Word Cloud}
\usage{
make_wordcloud(file, num_words = 100)
}
\arguments{
\item{file}{A .csv file of clean tweets}

\item{num_words}{The number of words in the word cloud, default is 100}
}
\description{
This function allows you to make a word cloud given a .csv file and the number of words for the word cloud.
}
\examples{
make_wordcloud()
}
\keyword{wordcloud}
