% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{n_Largest_coffee_producer}
\alias{n_Largest_coffee_producer}
\title{largest coffee producers}
\usage{
n_Largest_coffee_producer(n, data = coffee)
}
\arguments{
\item{number}{of  countries that contribute the most to coffee production}
}
\description{
\code{n_Largest_coffee_producer} returns the country and the region that is the largest coffee producer
}
\details{
The number for coffee production is product of two variables ; number of bags and bag weight
}
