% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_package.R
\name{new_package}
\alias{new_package}
\title{Create a package skeleton}
\usage{
new_package(
  name = "mypackage",
  dir = ".",
  author = getOption("coffee.name", "Joe Bloggs"),
  email = getOption("coffee.email", "Joe.Bloggs@missing.com"),
  enter = TRUE
)
}
\arguments{
\item{name}{Package name.}

\item{dir}{Directory to start in.}

\item{author}{Author.}

\item{email}{Email.}
}
\value{
Created directory (invisibly)
}
\description{
Create a package skeleton base on my preferred folder structure (hat tip to
Mark Van der Loo for Makefile idea)
}
