% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.R
\name{np}
\alias{np}
\title{Create basic package sans test setup}
\usage{
np(
  pkgname = "mypackage",
  dir,
  license = "MIT",
  firstname = getOption("FIRSTNAME", "Joe"),
  surname = getOption("SURNAME", "Bloggs"),
  email = getOption("EMAIL", "joe@bloggs.com"),
  orcid = getOption("ORCID"),
  open = TRUE,
  rstudio = TRUE
)
}
\arguments{
\item{pkgname}{The package name.}

\item{dir}{The directory to use for the package.}

\item{license}{What license to use for the package. Defaults to MIT.}

\item{firstname}{Your firstname.}

\item{surname}{Your surname.}

\item{email}{Your email.}

\item{orcid}{Your ORCID (optional).}

\item{open}{Open package directory after creation.}

\item{rstudio}{Use RStudio projects (.Rproj)?}
}
\description{
\code{np()} (new package) creates a minimal package with description.  It
borrows somewhat from Dirk Eddelbuettel's pkgKitten
\href{https://cran.r-project.org/web/packages/pkgKitten/index.html}{pkgKitten}
but changed to fit my workflow.
}
\details{
Authorship details will be grabbed from the options FIRSTNAME, SURNAME, EMAIL
and ORCID.
}
