% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot a colour palette}
\usage{
plot_palette(values, label = TRUE, square = FALSE)
}
\arguments{
\item{values}{character vector of named or hex colours.}

\item{label}{boolean. Do you want to label the plot or not? If \code{values} is
a named vector the names are used for labels, otherwise, the values.}

\item{square}{boolean. Display palette as square?}
}
\value{
The input \code{values} invisibly.
}
\description{
Plot a colour palette
}
\references{
Code inspired by hrbrmstr (https://stackoverflow.com/a/25726442)
}
