% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{Processing_method_Aroma}
\alias{Processing_method_Aroma}
\title{Difference between method of processing coffee and the coffee aroma}
\usage{
Processing_method_Aroma(data = coffee)
}
\description{
this function compares the total score for Aroma between different methods of preparation based on Tukey's honestly significant difference test
}
\details{
In order to normalize the data, outliers and the were removed

if the processing method was unknown, it was not included in the study.
}
\keyword{HSD}
\keyword{Tukey}
\keyword{linear}
\keyword{model,}
\keyword{regression}
\keyword{simple}
