% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{processing_method_totalscore_plot}
\alias{processing_method_totalscore_plot}
\title{A barplot of coffee total quality score grouped by the coffee color bean and the coffee processing method}
\usage{
processing_method_totalscore_plot(data = coffee)
}
\description{
A barplot of coffee total quality score grouped by the coffee color bean and the coffee processing method
}
