% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports_dang.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{lsos}
\alias{showMemoryUse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dang}{\code{\link[dang:ls.objects]{lsos}}, \code{\link[dang:ls.objects]{showMemoryUse}}}
}}

