% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Define Server logic for Coffee Preferences Shiny App}
\usage{
server(input, output)
}
\arguments{
\item{input}{Shiny input object that takes user selections for filters}

\item{output}{Shiny output object to render plots based on filtered data}
}
\value{
A Shiny server function

Reactive expression to filter data based on user inputs

This reactive function filters the data based on the selected age group, gender,
and favorite coffee. If "All" is selected for any input, the data is not filtered on that parameter.

A filtered data frame
Plot: Age Group Distribution by Gender

This plot shows the distribution of respondents by age group, with facets for male and female.

A ggplot object showing the count of respondents in each age group for each gender
Plot: Favorite Coffee by Age and Gender

This plot displays the count of respondents who prefer different coffee types, separated by gender.

A ggplot object showing the count of respondents for each favorite coffee by gender
}
\description{
The server function contains reactive expressions that filter the dataset based on user inputs for
age group, gender, and favorite coffee. It outputs two graphs: Age Group Distribution by Gender
and Favorite Coffee by Age and Gender.
}
