% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squares.R
\name{square}
\alias{square}
\title{A square function}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
The square of the input
}
\description{
This function allows you to square a variable.
}
