% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{strat.cleanup}
\alias{strat.cleanup}
\title{Remove .out files from a run}
\usage{
strat.cleanup(set = get("info"))
}
\arguments{
\item{set}{Detailed information of the current run, stored within this session's memory as variable \code{info}.}
}
\value{
NA
}
\description{
Remove all .out files in the folder of a strat file.
}
