% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_functions.R
\name{top20_producers_Aroma_sorted}
\alias{top20_producers_Aroma_sorted}
\title{top 20 countries form worst to best aroma}
\usage{
top20_producers_Aroma_sorted(data = coffee)
}
\description{
top 20 countries that are the largest coffee producers, sorted based on the score for coffee aroma
}
