% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{Define UI for Coffee Preferences Shiny App}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
A Shiny UI object
}
\description{
This UI includes three inputs for filtering the data: Age Group, Gender, and Favorite Coffee.
It outputs two bar plots: one for Age Group distribution by gender and another for Favorite Coffee distribution by age and gender.
}
\keyword{datasets}
