% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.R
\name{withinhpd}
\alias{withinhpd}
\title{Calculate the fit of a modelled age with the corresponding date}
\usage{
withinhpd(calibs, probs, modelled, prob = 0.95)
}
\arguments{
\item{calibs}{The calibrated ages of the distributions (dates)}

\item{probs}{The probabilities associated with the calibrated ages (dates)}

\item{modelled}{The modelled ages}

\item{prob}{Probability range for the hpd ranges}
}
\value{
a list of fits (values between 0 and 100%) for each date
}
\description{
This is a measure of the fit of the modelled age to that of the date. If many of the modelled age iterations fall within any of the highest posterior density (hpd) range of a date, the model fits the date well. The values can range from 0% (no fit, no modelled ages fall within any of the date's hpd ranges) to 100% (excellent fit).
}
