% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coffee}
\alias{coffee}
\title{Coffee Dataset}
\format{
A data frame with 989 rows and 22 variables
data(coffee)
}
\usage{
coffee
}
\description{
There is data for both Arabica and Robusta beans, across
countries and professionally rated. The two most
economically important varieties of coffee plant are the
Arabica and the Robusta; ~60\% of the coffee produced
worldwide is Arabica and ~40\% is Robusta. Arabica beans consist of 0.8–1.4\% caffeine and Robusta beans consist of 1.7–4\% caffeine. The Coffee Quality Institute is a non-profit organization that grades coffee samples from around the world in a consistent and professional manner. The result is published on their website and were scraped by reddit user JLD and posted on the forum’s dataset community. The dataset contains the grades the cuppers gave a sample on each of these ten attributes: Aroma, Flavor, Aftertaste, Acidity, Body, Balance, Uniformity, Clean Cup, Sweetness, Cupper Points. Each grade is on a 0–10 scale resulting to a total cupping score between zero and one hundred. Coffee cupping, or coffee tasting, is the practice of observing the tastes and aromas of brewed coffee. It is a professional practice but can be done informally by anyone or by professionals known as "Q Graders". A standard coffee cupping procedure involves deeply sniffing the coffee, then loudly slurping the coffee so it spreads to the back of the tongue. The coffee taster attempts to measure aspects of the coffee's taste, specifically the body (the texture or mouthfeel, such as oiliness), sweetness, acidity (a sharp and tangy feeling, like when biting into an orange), flavor (the characters in the cup), and aftertaste. Since coffee
beans embody telltale flavors from the region where they
were grown, cuppers may attempt to identify the coffee's origin.
}
\keyword{datasets}
