% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cat_graph.R
\name{make_cat_graph}
\alias{make_cat_graph}
\title{Making a Bar Graph for each Sentiment.}
\usage{
make_cat_graph(file)
}
\arguments{
\item{file}{A .csv file as a string of clean tweets}
}
\description{
This function allows you to analyze word counts that contribute to each sentiment.
}
\examples{
make_cat_graph()
}
\keyword{category}
\keyword{graph}
