% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot colours from names}
\usage{
plot_palette(values, label = TRUE, square = FALSE)
}
\arguments{
\item{values}{character vector of named or hex colours.}

\item{label}{boolean. Do you want to label the plot or not?}

\item{square}{boolean. Display palette as square?}
}
\description{
Plot colours from names
}
\references{
Code inspired by hrbrmstr (https://stackoverflow.com/a/25726442)
}
